requirements.txt
openai>=1.40
pydantic>=2.0
python-dotenv>=1.0
pandas>=2.1